requirements-minimal.txt
# ============================================================================
# Minimal Requirements for Agent V2 Core
# Only the essential dependencies needed to run the new agent architecture
# ============================================================================

# HTTP client (for LLM API calls)
requests>=2.31.0

# Git integration for repository analysis
GitPython>=3.1.40

# Web Search Tool (optional but recommended)
duckduckgo-search>=4.0.0

# Web content extraction (for web search tool)
trafilatura>=1.6.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# For evaluation scripts only
pandas>=2.0.0
scikit-learn>=1.3.0
tqdm>=4.66.0

# Installation
# ============================================================================
#
# Install minimal requirements:
#   pip install -r requirements-minimal.txt
#
# Prerequisites:
#   1. Install Ollama: https://ollama.ai
#   2. Pull a model: ollama pull codegemma:7b
#   3. Start Ollama: ollama serve
#
# Run agent:
#   python examples/basic_usage.py
#
# Run evaluation:
#   python src/evaluate_agent_v2.py --dataset transformed_data.csv
#
# ============================================================================
